Refactoring of Home/Screening/Login responsibility
