+++ EmployeeTempTracker/Controllers/LoginControllerLogic.cs
@@ -1,11 +1,6 @@
 using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
 using System.Text.Encodings.Web;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
 using EmployeeTempTracker.Models;
 
 namespace EmployeeTempTracker.Controllers {
@@ -18,7 +13,7 @@ namespace EmployeeTempTracker.Controllers {
         }
 
         // POST https://capstone.ohitski.org/Login/AuthUser
-        public IActionResult AuthUser(string uname, string passwd, int id = 1) {
+        public IActionResult AuthUser(string uname, string passwd, int id = 1) { // TODO: Move functionality to Login()?
             ViewData["Title"] = "Authenticate";
             // Handle user login API call here, unless it will be implemented in APIController.cs
             
@@ -47,5 +42,65 @@ namespace EmployeeTempTracker.Controllers {
             ViewData["Message"] = HtmlEncoder.Default.Encode($"Invalid username or password.");
             return View("InvalidLogin");
         }
+
+        // GET https://capstone.ohitski.org/Login/Login
+        public ActionResult Login(string DomainName, bool? SessionValid, string Username) {
+            LoginModel mod = new LoginModel();
+            mod.DomainName = DomainName;
+            if (SessionValid.HasValue)
+            {
+                mod.SessionValid = SessionValid.Value;
+                if (!SessionValid.Value)
+                {
+                    TempData["UserMessageErrorHeader"] = "Error";
+                    TempData["UserMessageErrorBody"] = "Invalid Session.  Please login.";
+                }
+            }
+            mod.Username = Username;
+            
+            return View(mod);
+        }
+
+        // POST https://capstone.ohitski.org/Login/Login
+        public IActionResult Login(LoginModel lm) {
+            int callerID = 117;
+            if (ModelState.IsValid)
+            {
+                //LoginModel am = new LoginModel();
+                AuthorizationModel am = new AuthorizationModel();
+                WsAuth.GXPAuthenticationSoapClient.EndpointConfiguration ec = WsAuth.GXPAuthenticationSoapClient.EndpointConfiguration.GXPAuthenticationSoap;
+                WsAuth.GXPAuthenticationSoapClient svc = new WsAuth.GXPAuthenticationSoapClient(ec);
+                WsAuth.LoginResult res = new WsAuth.LoginResult();
+                res = svc.Login(lm.DomainName, lm.Username, lm.Password, callerID);
+                //res.LoginSessionID;
+
+                if (String.IsNullOrEmpty(res.LoginSessionID))
+                {
+                    TempData["UserMessageErrorHeader"] = "Error";
+                    if (res.ServiceError != null)
+                    {
+                        TempData["UserMessageErrorBody"] = res.ServiceError.Message;
+                    }
+                    else
+                    {
+                        TempData["UserMessageErrorBody"] = "Login Failure";
+                    }
+                    // FormsAuthentication.SignOut();                   // does not exist
+                    return View("Login", lm);
+                }
+                else
+                {
+                    am.SessionID = res.LoginSessionID;
+                    am.UserName = lm.Username;
+                    am.Domain = lm.DomainName;
+                    return View("Login", lm);
+                }
+            }
+            else
+            {
+                return View("Login", lm);
+            }
+        }
+
     }
 }
