+++ EmployeeTempTracker/Controllers/ScreeningController.cs
@@ -1,5 +1,6 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
+using EmployeeTempTracker.Models;
 
 namespace EmployeeTempTracker.Controllers {
     public class ScreeningController : Controller {
@@ -13,5 +14,32 @@ namespace EmployeeTempTracker.Controllers {
         public IActionResult Index() {
             return viewProcessor_.Index();
         }
+
+        // GET https://capstone.ohitski.org/Screening/EnterScreening
+        public IActionResult EnterScreening() {
+            return viewProcessor_.EnterScreening();
+        }
+
+        // GET https://capstone.ohitski.org/Screening/ProcessScreening
+        public IActionResult ProcessScreening(string fname, string lname, string id, string org, string temperature, string symptoms, string closeContact, string intlTravel) {
+            return viewProcessor_.ProcessScreening(fname, lname, id, org, temperature, symptoms, closeContact, intlTravel);
+        }
+
+        // GET https://capstone.ohitski.org/Screening/SendHome
+        public IActionResult SendHome(ScreeningModel screening) {
+            return viewProcessor_.SendHome(screening);
+        }
+
+        // GET https://capstone.ohitski.org/Screening/ReviewScreening
+        public IActionResult ReviewScreening(ScreeningModel screening) { //TODO: instead of screening param here, pass Emp.Id
+            return viewProcessor_.ReviewScreening(screening);
+        }
+
+        // POST https://capstone.ohitski.org/Screening/Edit
+        [HttpPost]
+        public IActionResult Edit(ScreeningModel updatedScreening) {
+            return viewProcessor_.Edit(updatedScreening);
+        }
+
     }
 }
