+++ EmployeeTempTracker/Views/Home/Dashboard.cshtml
@@ -45,7 +45,7 @@
       </div>
       <div class="w3-third">
         <div class="w3-container">
-          <a href="/Home/EnterScreening" style="text-decoration: none;">
+          <a href="/Screening/EnterScreening" style="text-decoration: none;">
             <div class="w3-green">
               <h2 class="w3-center w3-padding-small">New Screening</h2>
             </div>
