+++ EmployeeTempTracker/Views/Login/Login.cshtml
@@ -1,0 +1,1 @@
+<p>Initial view to load in the LoginController Login() Functions</p>
