+++ EmployeeTempTracker/Views/Screening/EnterScreening.cshtml
@@ -6,7 +6,7 @@
 
 <div id="gsi_screening_form" class="w3-card-4 w3-margin-left w3-margin-right">
     <div class="w3-container w3-green"><h4 class="w3-center">New Screening</h4></div>
-    <form method="Post" action="/Home/ProcessScreening" class="w3-container">
+    <form method="Post" action="/Screening/ProcessScreening" class="w3-container">
         <!-- Begin Data Entry Portion -->
        
         <!-- First/Last Name Row -->
